Data structures for OxCOVID19 Database

Data structures for OxCOVID19 Database

data_structures

Format

A named list of 6 tibbles of data structures.

epidemiology

Tibble of description of fields for epidemiology table

government_response

Tibble of description of fields for government response table

country_statistics

Tibble of description for fields for country statistics table

mobility

Tibble of description for fields for mobility table

weather

Tibble of description for fields for weather table

administrative_division

Tibble of description for fields for administrative division table

Source

https://covid19.eng.ox.ac.uk/data_schema.html

Examples

data_structures

#> $epidemiology
#> # A tibble: 15 x 3
#>    Name          `Data type` Description                                        
#>    <chr>         <chr>       <chr>                                              
#>  1 SOURCE        VARCHAR     Data source                                        
#>  2 DATE          DATE        Day of the statistics                              
#>  3 GID           ARRAY       Unique geographical ID, for more details see  GADM 
#>  4 COUNTRY       VARCHAR     English name for the country                       
#>  5 COUNTRYCODE   VARCHAR(3)  ISO 3166-1 alpha-3 country codes                   
#>  6 ADM_AREA_1    VARCHAR     First-level administrative country subdivision     
#>  7 ADM_AREA_2    VARCHAR     Second-level administrative country subdivision    
#>  8 ADM_AREA_3    VARCHAR     Third-level administrative country subdivision     
#>  9 TESTED        INT         Number of people tested [cumulative]               
#> 10 CONFIRMED     INT         Number of confirmed cases [cumulative]             
#> 11 DEAD          INT         Number of deaths [cumulative]                      
#> 12 RECOVERED     INT         Number of confirmed who recovered [cumulative]     
#> 13 HOSPITALISED  INT         Number of confirmed who are/have been hospitalised…
#> 14 HOSPITALISED… INT         Number of confirmed who are/have been in the inten…
#> 15 QUARANTINED   INT         Number of confirmed with home quarantine [cumulati…
#> 
#> $government_response
#> # A tibble: 47 x 3
#>    Name              `Data type` Description                                    
#>    <chr>             <chr>       <chr>                                          
#>  1 SOURCE            VARCHAR     Data source                                    
#>  2 DATE              DATE        Day of the statistics                          
#>  3 GID               ARRAY       Unique geographical ID, for more details see  …
#>  4 COUNTRY           VARCHAR     English name for the country                   
#>  5 COUNTRYCODE       VARCHAR(3)  ISO 3166-1 alpha-3 country codes               
#>  6 ADM_AREA_1        VARCHAR     First-level administrative country subdivision 
#>  7 ADM_AREA_2        VARCHAR     Second-level administrative country subdivision
#>  8 ADM_AREA_3        VARCHAR     Third-level administrative country subdivision 
#>  9 ADM_AREA_3        VARCHAR     Third-level administrative country subdivision 
#> 10 C1_SCHOOL_CLOSING INT         Record closings of schools and universities    
#> # … with 37 more rows
#> 
#> $country_statistics
#> # A tibble: 8 x 3
#>   Name       `Data type` Description                                            
#>   <chr>      <chr>       <chr>                                                  
#> 1 SOURCE     VARCHAR     Data source                                            
#> 2 YEAR       INT         Day of the statistics                                  
#> 3 GID        ARRAY       Unique geographical ID, for more details see  GADM     
#> 4 COUNTRY    VARCHAR     English name for the country                           
#> 5 COUNTRYCO… VARCHAR(3)  ISO 3166-1 alpha-3 country codes                       
#> 6 ADM_LEVEL  VARCHAR     Aggregation level, 0 for World Bank and Integrated Val…
#> 7 SAMPLESIZE VARCHAR     Number of questions for Integrated Value Survey, -1 fo…
#> 8 PROPERTIES VARCHAR     Dictionary containing the region/country statistics. F…
#> 
#> $mobility
#> # A tibble: 17 x 3
#>    Name          `Data type` Description                                        
#>    <chr>         <chr>       <chr>                                              
#>  1 SOURCE        VARCHAR     Data source                                        
#>  2 DATE          DATE        Day of the statistics                              
#>  3 GID           ARRAY       Unique geographical ID, for more details see  GADM 
#>  4 COUNTRY       VARCHAR     English name for the country                       
#>  5 COUNTRYCODE   VARCHAR(3)  ISO 3166-1 alpha-3 country codes                   
#>  6 ADM_AREA_1    VARCHAR     First-level administrative country subdivision     
#>  7 ADM_AREA_2    VARCHAR     Second-level administrative country subdivision    
#>  8 ADM_AREA_3    VARCHAR     Third-level administrative country subdivision     
#>  9 TRANSIT_STAT… FLOAT       Mobility trends reported by Google for places like…
#> 10 RESIDENTIAL   FLOAT       Mobility trends reported by Google for places of r…
#> 11 WORKPLACE     FLOAT       Mobility trends reported by Google for places of w…
#> 12 PARKS         FLOAT       Mobility trends reported by Google for places like…
#> 13 RETAIL_RECRE… FLOAT       Mobility trends reported Google for places like re…
#> 14 GROCERY_FARM… FLOAT       Mobility trends reported by Google for places like…
#> 15 TRANSIT       FLOAT       The change in volume reported by Apple of people t…
#> 16 WALKING       FLOAT       The change in volume reported by Apple of people w…
#> 17 DRIVING       FLOAT       The change in volume reported by Apple of people d…
#> 
#> $weather
#> # A tibble: 36 x 3
#>    Name                 `Data type` Description                                 
#>    <chr>                <chr>       <chr>                                       
#>  1 DATE                 DATE        Day of the statistics                       
#>  2 GID                  VARCHAR     Unique geographical ID, for more details se…
#>  3 COUNTRY              VARCHAR     English name for the country                
#>  4 COUNTRYCODE          VARCHAR(3)  ISO 3166-1 alpha-3 country codes            
#>  5 PRECIP_MAX_AVG       FLOAT       Average of the daily maximum precipitation  
#>  6 PRECIP_MAX_STD       FLOAT       Standard deviation of the daily maximum pre…
#>  7 PRECIP_MEAN_AVG      FLOAT       Average of the daily mean precipitation     
#>  8 PRECIP_MEAN_STD      FLOAT       Standard deviation of the daily mean precip…
#>  9 SPECIFIC_HUMIDITY_M… FLOAT       Average of the daily maximum specific humid…
#> 10 SPECIFIC_HUMIDITY_M… FLOAT       Standard deviation of the daily maximum spe…
#> # … with 26 more rows
#> 
#> $administrative_division
#> # A tibble: 15 x 3
#>    Name              `Data type` Description                                    
#>    <chr>             <chr>       <chr>                                          
#>  1 GID               VARCHAR     Unique geographical ID, for more details see  …
#>  2 COUNTRY           VARCHAR     English name for the country                   
#>  3 COUNTRYCODE       VARCHAR(3)  ISO 3166-1 alpha-3 country codes               
#>  4 COUNTRYCODE_ALPH… VARCHAR     ISO 3166-1 alpha-2 country codes               
#>  5 ADM_AREA_1        VARCHAR     First-level administrative country subdivision 
#>  6 ADM_AREA_1_CODE   VARCHAR     First-level administrative country code subdiv…
#>  7 ADM_AREA_2        VARCHAR     Second-level administrative country subdivision
#>  8 ADM_AREA_2_CODE   VARCHAR     Second-level administrative country code subdi…
#>  9 ADM_AREA_3        VARCHAR     Third-level administrative country subdivision 
#> 10 ADM_AREA_3_CODE   VARCHAR     Third-level administrative country code subdiv…
#> 11 ADM_LEVEL         INTEGER     0 - for countries level, 1 - for regions etc.  
#> 12 LATITUDE          FLOAT       Geographic coordinate of region's centroid     
#> 13 LONGTITUDE        FLOAT       Geographic coordinate of region's centroid     
#> 14 PROPERTIES        JSON        Additional attributes describing region        
#> 15 GEOMETRY          GEOMETRY    Polygon describing geographical area           
#>