Get government response data based on specified query parameters

Get government response data based on specified query parameters

get_data_response(ccode = NULL, .start = NULL, .end = Sys.Date(), adm = NULL)

Arguments

ccode
Three character ISO country code for country to which data is required. Defaults to NULL which returns data for all countries.

.start
Starting date in YYYY-MM-DD character format of data to retrieve. Defaults to NULL which returns data from the earliest date available.

.end
End date in YYYY-MM-DD character format of data to retrieve. Defaults to current date as speciifed by Sys.Date.

adm
Numeric value for administrative level data required. Possible values are 0 for whole country, 1 for administrative level 1, 2 for administrative level 2 and 3 for administrative level 3. Default is NULL which returns data for all administrative divisions.

Value

A tibble of government response dataset based on specified parameters. If all parameters are kept to default, output is the same as that when get_table is used with tbl_name specified as government_response. However, this function will retrieve the full data into a local tibble which will take considerable time. It is therefore recommended that this function be used when the specific data requirements are already clearly determined. For data exploration, get_table shouild be used instead.

Examples

## Get government response data for the the whole of the UK for all
## available dates
get_data_response(ccode = "GBR",
                  adm = 0)

#> # A tibble: 234 x 46
#>    source date       gid   country countrycode adm_area_1 adm_area_2 adm_area_3
#>    <chr>  <date>     <chr> <chr>   <chr>       <chr>      <chr>      <chr>     
#>  1 GOVTR… 2020-01-01 GBR   United… GBR         NA         NA         NA        
#>  2 GOVTR… 2020-01-02 GBR   United… GBR         NA         NA         NA        
#>  3 GOVTR… 2020-01-03 GBR   United… GBR         NA         NA         NA        
#>  4 GOVTR… 2020-01-04 GBR   United… GBR         NA         NA         NA        
#>  5 GOVTR… 2020-01-05 GBR   United… GBR         NA         NA         NA        
#>  6 GOVTR… 2020-01-06 GBR   United… GBR         NA         NA         NA        
#>  7 GOVTR… 2020-01-07 GBR   United… GBR         NA         NA         NA        
#>  8 GOVTR… 2020-01-08 GBR   United… GBR         NA         NA         NA        
#>  9 GOVTR… 2020-01-09 GBR   United… GBR         NA         NA         NA        
#> 10 GOVTR… 2020-01-10 GBR   United… GBR         NA         NA         NA        
#> # … with 224 more rows, and 38 more variables: c1_school_closing <int>,
#> #   c1_flag <int>, c2_workplace_closing <int>, c2_flag <int>,
#> #   c3_cancel_public_events <int>, c3_flag <int>,
#> #   c4_restrictions_on_gatherings <int>, c4_flag <int>,
#> #   c5_close_public_transport <int>, c5_flag <int>,
#> #   c6_stay_at_home_requirements <int>, c6_flag <int>,
#> #   c7_restrictions_on_internal_movement <int>, c7_flag <int>,
#> #   c8_international_travel_controls <int>, e1_income_support <int>,
#> #   e1_flag <int>, e2_debtcontract_relief <int>, e3_fiscal_measures <dbl>,
#> #   e4_international_support <dbl>, h1_public_information_campaigns <int>,
#> #   h1_flag <int>, h2_testing_policy <int>, h3_contact_tracing <int>,
#> #   h4_emergency_investment_in_healthcare <dbl>,
#> #   h5_investment_in_vaccines <dbl>, m1_wildcard <chr>, stringency_index <dbl>,
#> #   stringency_indexfordisplay <dbl>, stringency_legacy_index <dbl>,
#> #   stringency_legacy_indexfordisplay <dbl>, government_response_index <dbl>,
#> #   government_response_index_for_display <dbl>,
#> #   containment_health_index <dbl>, containment_health_index_for_display <dbl>,
#> #   economic_support_index <dbl>, economic_support_index_for_display <dbl>,
#> #   actions <pq_jsonb>