Get weather data based on specified query parameters

Get weather data based on specified query parameters

get_data_weather(ccode = NULL, .start = NULL, .end = Sys.Date(), adm = NULL)

Arguments

ccode
Three character ISO country code for country to which data is required. Defaults to NULL which returns data for all countries.

.start
Starting date in YYYY-MM-DD character format of data to retrieve. Defaults to NULL which returns data from the earliest date available.

.end
End date in YYYY-MM-DD character format of data to retrieve. Defaults to current date as speciifed by Sys.Date.

adm
Numeric value for administrative level data required. Possible values are 0 for whole country, 1 for administrative level 1, 2 for administrative level 2 and 3 for administrative level 3. Default is NULL which returns data for all administrative divisions.

Value

A tibble of weather dataset based on specified parameters. If all parameters are kept to default, output is the same as that when get_table is used with tbl_name specified as weather. However, this function will retrieve the full data into a local tibble which will take considerable time. It is therefore recommended that this function be used when the specific data requirements are already clearly determined. For data exploration, get_table shouild be used instead.

Author

Ernest Guevarra

Examples

## Get weather data for the the whole of the UK for all available dates
get_data_weather(ccode = "GBR",
                 adm = 0)

#> # A tibble: 0 x 56
#> # … with 56 variables: source <chr>, date <date>, gid <chr>, country <chr>,
#> #   countrycode <chr>, adm_area_1 <chr>, adm_area_2 <chr>, adm_area_3 <chr>,
#> #   samplesize <int>, precipitation_max_avg <dbl>, precipitation_max_std <dbl>,
#> #   precipitation_mean_avg <dbl>, precipitation_mean_std <dbl>,
#> #   humidity_max_avg <dbl>, humidity_max_std <dbl>, humidity_mean_avg <dbl>,
#> #   humidity_mean_std <dbl>, humidity_min_avg <dbl>, humidity_min_std <dbl>,
#> #   sunshine_max_avg <dbl>, sunshine_max_std <dbl>, sunshine_mean_avg <dbl>,
#> #   sunshine_mean_std <dbl>, temperature_max_avg <dbl>,
#> #   temperature_max_std <dbl>, temperature_mean_avg <dbl>,
#> #   temperature_mean_std <dbl>, temperature_min_avg <dbl>,
#> #   temperature_min_std <dbl>, windgust_max_avg <dbl>, windgust_max_std <dbl>,
#> #   windgust_mean_avg <dbl>, windgust_mean_std <dbl>, windgust_min_avg <dbl>,
#> #   windgust_min_std <dbl>, windspeed_max_avg <dbl>, windspeed_max_std <dbl>,
#> #   windspeed_mean_avg <dbl>, windspeed_mean_std <dbl>,
#> #   windspeed_min_avg <dbl>, windspeed_min_std <dbl>,
#> #   cloudaltitude_max_valid <dbl>, cloudaltitude_max_avg <dbl>,
#> #   cloudaltitude_max_std <dbl>, cloudaltitude_min_valid <dbl>,
#> #   cloudaltitude_min_avg <dbl>, cloudaltitude_min_std <dbl>,
#> #   cloudaltitude_mean_valid <dbl>, cloudaltitude_mean_avg <dbl>,
#> #   cloudaltitude_mean_std <dbl>, cloudfrac_max_avg <dbl>,
#> #   cloudfrac_max_std <dbl>, cloudfrac_min_avg <dbl>, cloudfrac_min_std <dbl>,
#> #   cloudfrac_mean_avg <dbl>, cloudfrac_mean_std <dbl>