List out field names of specific tables available from OxCOVID19 Database

List out field names of specific tables available from OxCOVID19 Database

list_fields(con = connect_oxcovid19(), tab = list_tables())

Arguments

con
A PqConnection class object specifying the PostgreSQL connection to OxCOVID19 Database. This is produced by connect_covid19 function

tab
A character vector specifying table name/s to extract field names from. Must be a value included when a call to list_names is issued. Default is character vector of all table names specified by list_names.

Value

A named list of field names per specified table available from OxCOVID19 Database

Examples

list_fields()

#> $mobility
#>  [1] "source"            "date"              "country"          
#>  [4] "countrycode"       "adm_area_1"        "adm_area_2"       
#>  [7] "adm_area_3"        "transit_stations"  "residential"      
#> [10] "workplace"         "parks"             "retail_recreation"
#> [13] "grocery_pharmacy"  "gid"               "transit"          
#> [16] "walking"           "driving"          
#> 
#> $epidemiology
#>  [1] "source"           "date"             "country"          "countrycode"     
#>  [5] "adm_area_1"       "adm_area_2"       "adm_area_3"       "tested"          
#>  [9] "confirmed"        "recovered"        "dead"             "hospitalised"    
#> [13] "hospitalised_icu" "quarantined"      "gid"             
#> 
#> $government_response
#>  [1] "source"                               
#>  [2] "date"                                 
#>  [3] "gid"                                  
#>  [4] "country"                              
#>  [5] "countrycode"                          
#>  [6] "adm_area_1"                           
#>  [7] "adm_area_2"                           
#>  [8] "adm_area_3"                           
#>  [9] "c1_school_closing"                    
#> [10] "c1_flag"                              
#> [11] "c2_workplace_closing"                 
#> [12] "c2_flag"                              
#> [13] "c3_cancel_public_events"              
#> [14] "c3_flag"                              
#> [15] "c4_restrictions_on_gatherings"        
#> [16] "c4_flag"                              
#> [17] "c5_close_public_transport"            
#> [18] "c5_flag"                              
#> [19] "c6_stay_at_home_requirements"         
#> [20] "c6_flag"                              
#> [21] "c7_restrictions_on_internal_movement" 
#> [22] "c7_flag"                              
#> [23] "c8_international_travel_controls"     
#> [24] "e1_income_support"                    
#> [25] "e1_flag"                              
#> [26] "e2_debtcontract_relief"               
#> [27] "e3_fiscal_measures"                   
#> [28] "e4_international_support"             
#> [29] "h1_public_information_campaigns"      
#> [30] "h1_flag"                              
#> [31] "h2_testing_policy"                    
#> [32] "h3_contact_tracing"                   
#> [33] "h4_emergency_investment_in_healthcare"
#> [34] "h5_investment_in_vaccines"            
#> [35] "m1_wildcard"                          
#> [36] "stringency_index"                     
#> [37] "stringency_indexfordisplay"           
#> [38] "stringency_legacy_index"              
#> [39] "stringency_legacy_indexfordisplay"    
#> [40] "government_response_index"            
#> [41] "government_response_index_for_display"
#> [42] "containment_health_index"             
#> [43] "containment_health_index_for_display" 
#> [44] "economic_support_index"               
#> [45] "economic_support_index_for_display"   
#> [46] "actions"                              
#> 
#> $diagnostics
#> [1] "table_name"         "source"             "validation_success"
#> [4] "error"              "last_run_start"     "last_run_stop"     
#> [7] "first_timestamp"    "last_timestamp"    
#> 
#> $baseline_mortality
#> [1] "source"      "date"        "country"     "countrycode" "adm_level"  
#> [6] "gid"         "deaths"     
#> 
#> $weather
#>  [1] "date"                          "countrycode"                  
#>  [3] "gid"                           "precip_max_avg"               
#>  [5] "precip_max_std"                "precip_mean_avg"              
#>  [7] "precip_mean_std"               "specific_humidity_max_avg"    
#>  [9] "specific_humidity_max_std"     "specific_humidity_mean_avg"   
#> [11] "specific_humidity_mean_std"    "specific_humidity_min_avg"    
#> [13] "specific_humidity_min_std"     "short_wave_radiation_max_avg" 
#> [15] "short_wave_radiation_max_std"  "short_wave_radiation_mean_avg"
#> [17] "short_wave_radiation_mean_std" "air_temperature_max_avg"      
#> [19] "air_temperature_max_std"       "air_temperature_mean_avg"     
#> [21] "air_temperature_mean_std"      "air_temperature_min_avg"      
#> [23] "air_temperature_min_std"       "windgust_max_avg"             
#> [25] "windgust_max_std"              "windgust_mean_avg"            
#> [27] "windgust_mean_std"             "windgust_min_avg"             
#> [29] "windgust_min_std"              "windspeed_max_avg"            
#> [31] "windspeed_max_std"             "windspeed_mean_avg"           
#> [33] "windspeed_mean_std"            "windspeed_min_avg"            
#> [35] "windspeed_min_std"            
#> 
#> $administrative_division
#>  [1] "country"            "countrycode"        "countrycode_alpha2"
#>  [4] "adm_level"          "adm_area_1"         "adm_area_1_code"   
#>  [7] "adm_area_2"         "adm_area_2_code"    "adm_area_3"        
#> [10] "adm_area_3_code"    "gid"                "latitude"          
#> [13] "longitude"          "properties"         "geometry"          
#> 
#> $spatial_ref_sys
#> [1] "srid"      "auth_name" "auth_srid" "srtext"    "proj4text"
#> 
#> $geography_columns
#> [1] "f_table_catalog"    "f_table_schema"     "f_table_name"      
#> [4] "f_geography_column" "coord_dimension"    "srid"              
#> [7] "type"              
#> 
#> $geometry_columns
#> [1] "f_table_catalog"   "f_table_schema"    "f_table_name"     
#> [4] "f_geometry_column" "coord_dimension"   "srid"             
#> [7] "type"             
#> 
#> $country_statistics
#> [1] "source"      "year"        "country"     "countrycode" "adm_level"  
#> [6] "gid"         "samplesize"  "properties" 
#>