List out field names of specific tables available from OxCOVID19 Database

List out field names of specific tables available from OxCOVID19 Database

list_fields(con = connect_oxcovid19(), tab = list_tables())

Arguments

con
A PqConnection class object specifying the PostgreSQL connection to OxCOVID19 Database. This is produced by connect_covid19 function

tab
A character vector specifying table name/s to extract field names from. Must be a value included when a call to list_names is issued. Default is character vector of all table names specified by list_names.

Value

A named list of field names per specified table available from OxCOVID19 Database

Examples

list_fields()

#> $weather
#>  [1] "source"                   "date"                    
#>  [3] "gid"                      "country"                 
#>  [5] "countrycode"              "adm_area_1"              
#>  [7] "adm_area_2"               "adm_area_3"              
#>  [9] "samplesize"               "precipitation_max_avg"   
#> [11] "precipitation_max_std"    "precipitation_mean_avg"  
#> [13] "precipitation_mean_std"   "humidity_max_avg"        
#> [15] "humidity_max_std"         "humidity_mean_avg"       
#> [17] "humidity_mean_std"        "humidity_min_avg"        
#> [19] "humidity_min_std"         "sunshine_max_avg"        
#> [21] "sunshine_max_std"         "sunshine_mean_avg"       
#> [23] "sunshine_mean_std"        "temperature_max_avg"     
#> [25] "temperature_max_std"      "temperature_mean_avg"    
#> [27] "temperature_mean_std"     "temperature_min_avg"     
#> [29] "temperature_min_std"      "windgust_max_avg"        
#> [31] "windgust_max_std"         "windgust_mean_avg"       
#> [33] "windgust_mean_std"        "windgust_min_avg"        
#> [35] "windgust_min_std"         "windspeed_max_avg"       
#> [37] "windspeed_max_std"        "windspeed_mean_avg"      
#> [39] "windspeed_mean_std"       "windspeed_min_avg"       
#> [41] "windspeed_min_std"        "cloudaltitude_max_valid" 
#> [43] "cloudaltitude_max_avg"    "cloudaltitude_max_std"   
#> [45] "cloudaltitude_min_valid"  "cloudaltitude_min_avg"   
#> [47] "cloudaltitude_min_std"    "cloudaltitude_mean_valid"
#> [49] "cloudaltitude_mean_avg"   "cloudaltitude_mean_std"  
#> [51] "cloudfrac_max_avg"        "cloudfrac_max_std"       
#> [53] "cloudfrac_min_avg"        "cloudfrac_min_std"       
#> [55] "cloudfrac_mean_avg"       "cloudfrac_mean_std"      
#> 
#> $world_bank
#>  [1] "source"         "gid"            "country_name"   "country_code"  
#>  [5] "adm_area_1"     "adm_area_2"     "adm_area_3"     "indicator_name"
#>  [9] "indicator_code" "value"          "year"          
#> 
#> $mobility
#>  [1] "source"            "date"              "country"          
#>  [4] "countrycode"       "adm_area_1"        "adm_area_2"       
#>  [7] "adm_area_3"        "transit_stations"  "residential"      
#> [10] "workplace"         "parks"             "retail_recreation"
#> [13] "grocery_pharmacy"  "gid"               "transit"          
#> [16] "walking"           "driving"          
#> 
#> $epidemiology
#>  [1] "source"           "date"             "country"          "countrycode"     
#>  [5] "adm_area_1"       "adm_area_2"       "adm_area_3"       "tested"          
#>  [9] "confirmed"        "recovered"        "dead"             "hospitalised"    
#> [13] "hospitalised_icu" "quarantined"      "gid"             
#> 
#> $government_response
#>  [1] "source"                               
#>  [2] "date"                                 
#>  [3] "gid"                                  
#>  [4] "country"                              
#>  [5] "countrycode"                          
#>  [6] "adm_area_1"                           
#>  [7] "adm_area_2"                           
#>  [8] "adm_area_3"                           
#>  [9] "c1_school_closing"                    
#> [10] "c1_flag"                              
#> [11] "c2_workplace_closing"                 
#> [12] "c2_flag"                              
#> [13] "c3_cancel_public_events"              
#> [14] "c3_flag"                              
#> [15] "c4_restrictions_on_gatherings"        
#> [16] "c4_flag"                              
#> [17] "c5_close_public_transport"            
#> [18] "c5_flag"                              
#> [19] "c6_stay_at_home_requirements"         
#> [20] "c6_flag"                              
#> [21] "c7_restrictions_on_internal_movement" 
#> [22] "c7_flag"                              
#> [23] "c8_international_travel_controls"     
#> [24] "e1_income_support"                    
#> [25] "e1_flag"                              
#> [26] "e2_debtcontract_relief"               
#> [27] "e3_fiscal_measures"                   
#> [28] "e4_international_support"             
#> [29] "h1_public_information_campaigns"      
#> [30] "h1_flag"                              
#> [31] "h2_testing_policy"                    
#> [32] "h3_contact_tracing"                   
#> [33] "h4_emergency_investment_in_healthcare"
#> [34] "h5_investment_in_vaccines"            
#> [35] "m1_wildcard"                          
#> [36] "stringency_index"                     
#> [37] "stringency_indexfordisplay"           
#> [38] "stringency_legacy_index"              
#> [39] "stringency_legacy_indexfordisplay"    
#> [40] "government_response_index"            
#> [41] "government_response_index_for_display"
#> [42] "containment_health_index"             
#> [43] "containment_health_index_for_display" 
#> [44] "economic_support_index"               
#> [45] "economic_support_index_for_display"   
#> [46] "actions"                              
#> 
#> $world_bank_time_series
#>  [1] "source"         "gid"            "country_name"   "country_code"  
#>  [5] "adm_area_1"     "adm_area_2"     "adm_area_3"     "indicator_name"
#>  [9] "indicator_code" "1960"           "1961"           "1962"          
#> [13] "1963"           "1964"           "1965"           "1966"          
#> [17] "1967"           "1968"           "1969"           "1970"          
#> [21] "1971"           "1972"           "1973"           "1974"          
#> [25] "1975"           "1976"           "1977"           "1978"          
#> [29] "1979"           "1980"           "1981"           "1982"          
#> [33] "1983"           "1984"           "1985"           "1986"          
#> [37] "1987"           "1988"           "1989"           "1990"          
#> [41] "1991"           "1992"           "1993"           "1994"          
#> [45] "1995"           "1996"           "1997"           "1998"          
#> [49] "1999"           "2000"           "2001"           "2002"          
#> [53] "2003"           "2004"           "2005"           "2006"          
#> [57] "2007"           "2008"           "2009"           "2010"          
#> [61] "2011"           "2012"           "2013"           "2014"          
#> [65] "2015"           "2016"           "2017"           "2018"          
#> [69] "2019"          
#> 
#> $diagnostics
#> [1] "table_name"         "source"             "validation_success"
#> [4] "error"              "last_run_start"     "last_run_stop"     
#> [7] "first_timestamp"    "last_timestamp"    
#> 
#> $baseline_mortality
#> [1] "source"      "date"        "country"     "countrycode" "adm_level"  
#> [6] "gid"         "deaths"     
#> 
#> $administrative_division
#>  [1] "country"            "countrycode"        "countrycode_alpha2"
#>  [4] "adm_level"          "adm_area_1"         "adm_area_1_code"   
#>  [7] "adm_area_2"         "adm_area_2_code"    "adm_area_3"        
#> [10] "adm_area_3_code"    "gid"                "latitude"          
#> [13] "longitude"          "properties"         "geometry"          
#> 
#> $spatial_ref_sys
#> [1] "srid"      "auth_name" "auth_srid" "srtext"    "proj4text"
#> 
#> $geography_columns
#> [1] "f_table_catalog"    "f_table_schema"     "f_table_name"      
#> [4] "f_geography_column" "coord_dimension"    "srid"              
#> [7] "type"              
#> 
#> $geometry_columns
#> [1] "f_table_catalog"   "f_table_schema"    "f_table_name"     
#> [4] "f_geometry_column" "coord_dimension"   "srid"             
#> [7] "type"             
#> 
#> $country_statistics
#> [1] "source"      "year"        "country"     "countrycode" "adm_level"  
#> [6] "gid"         "samplesize"  "properties" 
#>